+++ var/flows/main/dependencies.js
@@ -2,6 +2,7 @@ define( [
    'laxar-uikit/controls/i18n/ax-i18n-control',
    'finder-box-control/finder-box-control',
    'finder-demo-utilities/finder-demo-utilities',
+   'laxar-developer-tools-widget/ax-developer-tools-widget',
    'laxar-application/includes/widgets/finder-demo/music-brainz-search-widget/music-brainz-search-widget',
    'laxar-application/includes/widgets/finder-demo/omdb-search-widget/omdb-search-widget',
    'laxar-application/includes/widgets/finder-demo/open-street-map-widget/open-street-map-widget',
@@ -15,7 +16,7 @@ define( [
 
    var modules = [].slice.call( arguments );
    return {
-      'angular': modules.slice( 0, 9 ),
-      'plain': modules.slice( 9, 11 )
+      'angular': modules.slice( 0, 10 ),
+      'plain': modules.slice( 10, 12 )
    };
 } );