+++ CHANGELOG.md
@@ -3,8 +3,11 @@
 ## Last Changes
 
 
-## v1.4.1
+## v1.5.0
 
+- [#17](https://github.com/LaxarJS/mashup-demo/issues/17): fixed copyright holder in license
+- [#16](https://github.com/LaxarJS/mashup-demo/issues/16): expanded widget definitions and updated laxar
+- [#15](https://github.com/LaxarJS/finder-demo/issues/15): fixed laxar-mocks/laxar dependency mismatch
 - [#14](https://github.com/LaxarJS/finder-demo/issues/14): fixed jasmine2 paths in spec-runners and require_config